Style

+++ database/migrations/2020_05_24_234029_add_punch_permissions.php
@@ -26,6 +26,8 @@ class AddPunchPermissions extends Migration
     public function down()
     {
         $dbPerm = Permission::where('name', 'record-punches')->first();
-        $dbPerm->delete();
+        if ($dbPerm !== null) {
+            $dbPerm->delete();
+        }
     }
 }
